requirements.txt
fastapi
uvicorn
requests
jinja2
python-dotenv
openai
beautifulsoup4
html5lib
playwright
easyocr
opencv-python
python-multipart
